assign ownership of spawned forms and it works
Was having problems passing a file-arg on the cl to extant instances of
Yata when either the font-picker or the clipboard-editor were open (and left
on top of the mainwindow). After much unsuccessful and curmudgeonly
p/invoking (hint: reading through and implementing 50+ threads on
StackOverflow) that merely giving ownership of the sub-forms to the
main-form solves it. Time to watch a vid about aLiens ... ie, a bit of well
deserved R&R after spending three days on a 30 second fix.