differ sync - sync selected cells/rows/cols ALWAYS
as well as being a convenient navigation device between diffed tables
it will hopefully prevent a scrollbar anomaly where, if the diffed tables
have different selections, and either of the selections is not currently
displayed on its table, and the form is resized, the scrollbar values no
longer match the display of the table (bork!).

I believe that is caused by EnsureDisplayed() fighting with scroll-syncing

- usually handled by YataForm.SyncSelect() but rowhead and colhead clicks
get their own routinues.